tabele-v6-modular/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2

# Azure Services
azure-ai-formrecognizer==3.3.0
azure-ai-inference==1.0.0b1
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-core==1.29.5

# Template engine
jinja2==3.1.2

# Document processing (legacy - consider removing if not needed)
python-docx==1.1.0
PyPDF2==3.0.1